requirements.txt
lucide_fasthtml==0.0.9
python_fasthtml==0.6.14
